+++ src/card/cards/stale_followups.rs
@@ -124,3 +124,5 @@ pub const TOUCH_OF_DARKNESS: CardDefinitionId = CardDefinitionId(1768);
 pub const DWARVEN_SONG: CardDefinitionId = CardDefinitionId(1769);
 pub const SYLVAN_PARADISE: CardDefinitionId = CardDefinitionId(1770);
 pub const TWO_HEADED_GIANT_OF_FORIYS: CardDefinitionId = CardDefinitionId(1771);
+pub const BENALISH_HERO: CardDefinitionId = CardDefinitionId(1772);
+pub const MESA_PEGASUS: CardDefinitionId = CardDefinitionId(1773);